// app/api-client/models/StatCardResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Minarai Dojo API
 * API de gestión para Dojo de Karate - Pagos y Autenticación
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { StatCardResponseTrend } from './StatCardResponseTrend';
import {
    StatCardResponseTrendFromJSON,
    StatCardResponseTrendFromJSONTyped,
    StatCardResponseTrendToJSON,
    StatCardResponseTrendToJSONTyped,
} from './StatCardResponseTrend';

/**
 * 
 * @export
 * @interface StatCardResponse
 */
export interface StatCardResponse {
    /**
     * 
     * @type {number}
     * @memberof StatCardResponse
     */
    value?: number;
    /**
     * 
     * @type {StatCardResponseTrend}
     * @memberof StatCardResponse
     */
    trend?: StatCardResponseTrend;
}

/**
 * Check if a given object implements the StatCardResponse interface.
 */
export function instanceOfStatCardResponse(value: object): value is StatCardResponse {
    return true;
}

export function StatCardResponseFromJSON(json: any): StatCardResponse {
    return StatCardResponseFromJSONTyped(json, false);
}

export function StatCardResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): StatCardResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'value': json['value'] == null ? undefined : json['value'],
        'trend': json['trend'] == null ? undefined : StatCardResponseTrendFromJSON(json['trend']),
    };
}

export function StatCardResponseToJSON(json: any): StatCardResponse {
    return StatCardResponseToJSONTyped(json, false);
}

export function StatCardResponseToJSONTyped(value?: StatCardResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'value': value['value'],
        'trend': StatCardResponseTrendToJSON(value['trend']),
    };
}